move AppOptions, IOSAppOptions and AndroidAppOptions to src/crossplatform
