+++ packages/patrol_cli/lib/src/android/android_test_backend.dart
@@ -1,97 +1,15 @@
-import 'dart:convert' show base64Encode, utf8;
 import 'dart:io' show Process;
 
 import 'package:adb/adb.dart';
 import 'package:dispose_scope/dispose_scope.dart';
 import 'package:file/file.dart';
-import 'package:path/path.dart' show basename;
 import 'package:patrol_cli/src/base/exceptions.dart';
 import 'package:patrol_cli/src/base/logger.dart';
 import 'package:patrol_cli/src/base/process.dart';
+import 'package:patrol_cli/src/crossplatform/app_options.dart';
 import 'package:patrol_cli/src/devices.dart';
-import 'package:patrol_cli/src/features/test/test_backend.dart';
 import 'package:platform/platform.dart';
 import 'package:process/process.dart';
-
-class AndroidAppOptions extends AppOptions {
-  const AndroidAppOptions({
-    required super.target,
-    required super.flavor,
-    required super.dartDefines,
-    this.packageName,
-  });
-
-  final String? packageName;
-
-  @override
-  String get description => 'apk with entrypoint ${basename(target)}';
-
-  List<String> toGradleAssembleInvocation({required bool isWindows}) {
-    return _toGradleInvocation(
-      isWindows: isWindows,
-      task: 'assemble${_effectiveFlavor}Debug',
-    );
-  }
-
-  List<String> toGradleAssembleTestInvocation({required bool isWindows}) {
-    return _toGradleInvocation(
-      isWindows: isWindows,
-      task: 'assemble${_effectiveFlavor}DebugAndroidTest',
-    );
-  }
-
-  List<String> toGradleConnectedTestInvocation({required bool isWindows}) {
-    return _toGradleInvocation(
-      isWindows: isWindows,
-      task: 'connected${_effectiveFlavor}DebugAndroidTest',
-    );
-  }
-
-  String get _effectiveFlavor {
-    var flavor = this.flavor ?? '';
-    if (flavor.isNotEmpty) {
-      flavor = flavor[0].toUpperCase() + flavor.substring(1);
-    }
-    return flavor;
-  }
-
-  /// Translates these options into a proper Gradle invocation.
-  List<String> _toGradleInvocation({
-    required bool isWindows,
-    required String task,
-  }) {
-    final List<String> cmd;
-    if (isWindows) {
-      cmd = <String>[r'.\gradlew.bat'];
-    } else {
-      cmd = <String>['./gradlew'];
-    }
-
-    // Add Gradle task
-    cmd.add(':app:$task');
-
-    // Add Dart test target
-    final target = '-Ptarget=${this.target}';
-    cmd.add(target);
-
-    // Add Dart defines encoded in base64
-    if (dartDefines.isNotEmpty) {
-      final dartDefinesString = StringBuffer();
-      for (var i = 0; i < dartDefines.length; i++) {
-        final entry = dartDefines.entries.elementAt(i);
-        final pair = utf8.encode('${entry.key}=${entry.value}');
-        dartDefinesString.write(base64Encode(pair));
-        if (i != dartDefines.length - 1) {
-          dartDefinesString.write(',');
-        }
-      }
-
-      cmd.add('-Pdart-defines=$dartDefinesString');
-    }
-
-    return cmd;
-  }
-}
 
 /// Provides functionality to build, install, run, and uninstall Android apps.
 ///
@@ -224,7 +142,6 @@ class AndroidTestBackend {
     });
   }
 
-  @override
   Future<void> uninstall(String appId, Device device) async {
     _logger.detail('Uninstalling $appId from ${device.name}');
     await _adb.uninstall(appId, device: device.id);
