+++ packages/patrol_cli/test/fakes.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:mocktail/mocktail.dart';
-import 'package:patrol_cli/src/android/android_test_backend.dart';
+import 'package:patrol_cli/src/crossplatform/app_options.dart';
 import 'package:patrol_cli/src/devices.dart';
 import 'package:patrol_cli/src/ios/ios_test_backend.dart';
